+++ CHANGELOG.md
@@ -14,6 +14,8 @@ These changes are available on the `master` branch, but have not yet been releas
 
 - Added `Member.colours` and `Member.colors` properties.
   ([#3063](https://github.com/Pycord-Development/pycord/pull/3063))
+- Added `Member.nameplate` to access a member's guild-specific nameplate.
+  ([#3104](https://github.com/Pycord-Development/pycord/pull/3104))
 
 ### Changed